@@ -5,9 +5,7 @@ concluded events.
 """
 
 import datetime
-from typing import Any, Collection, Dict, List, Optional, Set, Tuple, Union
-
-from typing_extensions import Protocol
+from typing import Any, Collection, Dict, List, Optional, Protocol, Set, Tuple, Union
 
 import cdedb.database.constants as const
 import cdedb.validationtypes as vtypes
@@ -15,6 +13,7 @@ from cdedb.backend.common import (
     AbstractBackend, Silencer, access, affirm_set_validation as affirm_set,
     affirm_validation_typed as affirm,
     affirm_validation_typed_optional as affirm_optional, singularize,
+    read_conditional_write_composer,
 )
 from cdedb.backend.event import EventBackend
 from cdedb.common import (
@@ -159,6 +158,12 @@ class PastEventBackend(AbstractBackend):
             self.past_event_log(rs, const.PastEventLogCodes.institution_changed,
                                 pevent_id=None, change_note=current['title'])
         return ret
+
+    class _RCWInstitutionProtocol(Protocol):
+        def __call__(self, rs: RequestState, data: CdEDBObject
+                     ) -> DefaultReturnCode: ...
+    rcw_institution: _RCWInstitutionProtocol = read_conditional_write_composer(
+        get_institution, set_institution, id_param_name='institution_id')
 
     @access("cde_admin", "event_admin")
     def create_institution(self, rs: RequestState, data: CdEDBObject
@@ -523,11 +528,12 @@ class PastEventBackend(AbstractBackend):
                     self.remove_participant(rs, pevent_id, pcourse_id=None,
                                             persona_id=persona_id)
                 else:
-                    rs.notify("error", n_("User is already pure event participant."))
                     return 0
-            ret = self.sql_insert(rs, "past_event.participants", data)
-            self.past_event_log(rs, const.PastEventLogCodes.participant_added,
-                                pevent_id, persona_id=persona_id)
+            ret = self.sql_insert(rs, "past_event.participants", data,
+                                  drop_on_conflict=True)
+            if ret:
+                self.past_event_log(rs, const.PastEventLogCodes.participant_added,
+                                    pevent_id, persona_id=persona_id)
         return ret
 
     @access("cde_admin", "event_admin")
@@ -734,9 +740,7 @@ class PastEventBackend(AbstractBackend):
             if course_id not in courses_seen:
                 self.delete_past_course(rs, course_map[course_id])
             elif not courses[course_id]['active_segments']:
-                self.logger.warning(
-                    "Course {} remains without active parts.".format(
-                        course_id))
+                self.logger.warning(f"Course {course_id} remains without active parts.")
         return new_id
 
     @access("cde_admin", "event_admin")
