@@ -333,6 +333,8 @@ _DEFAULTS = {
     "MAIL_DOMAIN": "db.cde-ev.de",
     # host to use for sending emails
     "MAIL_HOST": "localhost",
+    # email for internal system trouble notifications
+    "TROUBLESHOOTING_ADDRESS": "admin@cde-ev.de",
 
     # email for cde account requests
     "CDE_ADMIN_ADDRESS": "cde-admins@cde-ev.de",
@@ -344,7 +346,7 @@ _DEFAULTS = {
     "ASSEMBLY_ADMIN_ADDRESS": "vorstand@cde-ev.de",
 
     # email for privilege changes
-    "META_ADMIN_ADDRESS": "admin@lists.cde-ev.de",
+    "META_ADMIN_ADDRESS": "admin@cde-ev.de",
 
     # email for ballot tallies
     "BALLOT_TALLY_ADDRESS": "wahlbekanntmachung@lists.cde-ev.de",
@@ -637,7 +639,7 @@ class BasicConfig(Mapping[str, Any]):
     # noinspection PyUnresolvedReferences
     def __init__(self) -> None:
         try:
-            import cdedb.localconfig as config_mod
+            import cdedb.localconfig as config_mod  # pylint: disable=import-outside-toplevel
             config = {
                 key: getattr(config_mod, key)
                 for key in _BASIC_DEFAULTS.keys() & set(dir(config_mod))
@@ -672,7 +674,7 @@ class Config(BasicConfig):
         :param configpath: path to file with overrides
         """
         super().__init__()
-        _LOGGER.debug("Initialising Config with path {}".format(configpath))
+        _LOGGER.debug(f"Initialising Config with path {configpath}")
         self._configpath = configpath
         config_keys = _DEFAULTS.keys() | _BASIC_DEFAULTS.keys()
 
@@ -693,7 +695,7 @@ class Config(BasicConfig):
 
         try:
             # noinspection PyUnresolvedReferences
-            import cdedb.localconfig as secondaryconf_mod
+            import cdedb.localconfig as secondaryconf_mod  # pylint: disable=import-outside-toplevel
             secondaryconf = {
                 key: getattr(secondaryconf_mod, key)
                 for key in config_keys & set(dir(secondaryconf_mod))
@@ -736,7 +738,7 @@ class SecretsConfig(Mapping[str, Any]):
 
         try:
             # noinspection PyUnresolvedReferences
-            import cdedb.localconfig as secondaryconf_mod
+            import cdedb.localconfig as secondaryconf_mod  # pylint: disable=import-outside-toplevel
             secondaryconf = {
                 key: getattr(secondaryconf_mod, key)
                 for key in _SECRECTS_DEFAULTS.keys() & set(
