@@ -10,13 +10,12 @@ This should be the only module which makes subsistantial use of psycopg.
 
 import logging
 from types import TracebackType
-from typing import Any, Collection, Mapping, NoReturn, Optional, Type
+from typing import Any, Collection, Literal, Mapping, Protocol, NoReturn, Optional, Type
 
 import psycopg2
 import psycopg2.extensions
 import psycopg2.extras
 from psycopg2.extensions import ISOLATION_LEVEL_SERIALIZABLE as SERIALIZABLE
-from typing_extensions import Literal, Protocol
 
 # We cannot import cdedb.config here.
 # from cdedb.config import SecretsConfig
@@ -69,7 +68,7 @@ def _create_connection(dbname: str, dbuser: str, password: str, port: int,
         conn = psycopg2.connect(**connection_parameters, host="cdb", port=5432)
     conn.set_client_encoding("UTF8")
     conn.set_session(isolation_level)
-    _LOGGER.debug("Created connection to {} as {}".format(dbname, dbuser))
+    _LOGGER.debug(f"Created connection to {dbname} as {dbuser}")
     return conn
 
 
