@@ -6,28 +6,36 @@ organization. We will speak of members in most contexts where former
 members are also possible.
 """
 
+import copy
 import datetime
 import decimal
 from collections import OrderedDict
-from typing import Any, Collection, Dict, List, Optional, Tuple
+from typing import Any, Collection, Dict, List, Optional, Protocol, Tuple
 
-from typing_extensions import Protocol
+import psycopg2.extensions
 
 import cdedb.database.constants as const
 import cdedb.validationtypes as vtypes
 from cdedb.backend.common import (
     AbstractBackend, access, affirm_set_validation as affirm_set,
+    affirm_array_validation as affirm_array,
     affirm_validation_typed as affirm,
     affirm_validation_typed_optional as affirm_optional, batchify, singularize,
 )
+from cdedb.backend.past_event import PastEventBackend
 from cdedb.common import (
     EXPULS_PERIOD_FIELDS, LASTSCHRIFT_FIELDS, LASTSCHRIFT_TRANSACTION_FIELDS,
     ORG_PERIOD_FIELDS, CdEDBLog, CdEDBObject, CdEDBObjectMap, DefaultReturnCode,
     DeletionBlockers, PrivilegeError, QuotaException, RequestState, implying_realms,
-    merge_dicts, n_, now, unwrap,
+    merge_dicts, n_, now, unwrap, glue, LineResolutions, PathLike, make_proxy,
+    PARSE_OUTPUT_DATEFORMAT, ArchiveError,
 )
 from cdedb.database.connection import Atomizer
+from cdedb.filter import money_filter
 from cdedb.query import Query, QueryOperators, QueryScope
+from cdedb.validation import (
+    PERSONA_CDE_CREATION as CDE_TRANSITION_FIELDS, is_optional,
+)
 
 
 class CdEBackend(AbstractBackend):
@@ -36,6 +44,10 @@ class CdEBackend(AbstractBackend):
     .. note:: The changelog functionality is to be found in the core backend.
     """
     realm = "cde"
+
+    def __init__(self, configpath: PathLike = None):
+        super().__init__(configpath)
+        self.pastevent = make_proxy(PastEventBackend(configpath), internal=True)
 
     @classmethod
     def is_admin(cls, rs: RequestState) -> bool:
@@ -101,6 +113,56 @@ class CdEBackend(AbstractBackend):
             submitted_by=submitted_by, additional_columns=additional_columns,
             change_note=change_note, time_start=time_start,
             time_stop=time_stop)
+
+    @access("core_admin", "cde_admin")
+    def change_membership(
+            self, rs: RequestState, persona_id: int, is_member: bool
+    ) -> Tuple[DefaultReturnCode, List[int], bool]:
+        """Special modification function for membership.
+
+        This is similar to the version from the core backend, but can
+        additionally handle lastschrift permits.
+
+        In the general situation this variant should be used.
+
+        :param is_member: Desired target state of membership.
+        :returns: A tuple containing the return code, a list of ids of
+            revoked lastschrift permits and whether any in-flight
+            transactions are affected.
+        """
+        persona_id = affirm(vtypes.ID, persona_id)
+        is_member = affirm(bool, is_member)
+        code = 1
+        revoked_permits = []
+        collateral_transactions = False
+        with Atomizer(rs):
+            if not is_member:
+                lastschrift_ids = self.list_lastschrift(
+                    rs, persona_ids=(persona_id,), active=None)
+                lastschrifts = self.get_lastschrifts(
+                    rs, lastschrift_ids.keys())
+                active_permits = []
+                for lastschrift in lastschrifts.values():
+                    if not lastschrift['revoked_at']:
+                        active_permits.append(lastschrift['id'])
+                if active_permits:
+                    if self.list_lastschrift_transactions(
+                            rs, lastschrift_ids=active_permits,
+                            stati=(const.LastschriftTransactionStati.issued,)):
+                        collateral_transactions = True
+                    for active_permit in active_permits:
+                        data = {
+                            'id': active_permit,
+                            'revoked_at': now(),
+                        }
+                        lastschrift_code = self.set_lastschrift(rs, data)
+                        if lastschrift_code <= 0:
+                            raise ValueError(n_(
+                                "Failed to revoke active lastschrift permit"))
+                        revoked_permits.append(active_permit)
+            code = self.core.change_membership_easy_mode(
+                rs, persona_id, is_member)
+        return code, revoked_permits, collateral_transactions
 
     @access("member", "core_admin", "cde_admin")
     def list_lastschrift(self, rs: RequestState,
@@ -435,7 +497,7 @@ class CdEBackend(AbstractBackend):
                     const.FinanceLogCodes.lastschrift_transaction_success,
                     change_note="Erfolgreicher Lastschrifteinzug.")
                 if new_balance >= self.conf["MEMBERSHIP_FEE"]:
-                    self.core.change_membership(rs, persona_id, is_member=True)
+                    self.change_membership(rs, persona_id, is_member=True)
                 # Return early since change_persona_balance does the logging
                 return ret
             elif status == const.LastschriftTransactionStati.failure:
@@ -452,6 +514,20 @@ class CdEBackend(AbstractBackend):
             self.core.finance_log(rs, code, persona_id, delta, new_balance,
                                   change_note=str(update['tally']))
         return ret
+
+    @access("finance_admin")
+    def finalize_lastschrift_transactions(
+            self, rs: RequestState, transactions: List[CdEDBObject]
+    ) -> DefaultReturnCode:
+        """Atomized multiplex variant of finalize_lastschrift_transaction."""
+        transactions = affirm_array(vtypes.LastschriftTransactionEntry, transactions)
+        code = 1
+        with Atomizer(rs):
+            for transaction in transactions:
+                code *= self.finalize_lastschrift_transaction(
+                    rs, transaction['transaction_id'], transaction['status'],
+                    tally=transaction['tally'])
+        return code
 
     @access("finance_admin")
     def rollback_lastschrift_transaction(
@@ -555,6 +631,77 @@ class CdEBackend(AbstractBackend):
                 rs, const.FinanceLogCodes.lastschrift_transaction_skip,
                 lastschrift['persona_id'], None, None)
         return ret
+
+    @access("finance_admin")
+    def perform_money_transfers(self, rs: RequestState, data: List[CdEDBObject]
+                                ) -> Tuple[bool, Optional[int], Optional[int]]:
+        """Resolve all money transfer entries.
+
+        :returns: A bool indicating success and:
+            * In case of success:
+                * The number of recorded transactions
+                * The number of new members.
+            * In case of error:
+                * The index of the erronous line or None
+                    if a DB-serialization error occurred.
+                * None
+        """
+        data = affirm_array(vtypes.MoneyTransferEntry, data)
+        index = 0
+        note_template = ("Guthabenänderung um {amount} auf {new_balance} "
+                         "(Überwiesen am {date})")
+        # noinspection PyBroadException
+        try:
+            with Atomizer(rs):
+                count = 0
+                memberships_gained = 0
+                persona_ids = tuple(e['persona_id'] for e in data)
+                personas = self.core.get_total_personas(rs, persona_ids)
+                for index, datum in enumerate(data):
+                    assert isinstance(datum['amount'], decimal.Decimal)
+                    new_balance = (personas[datum['persona_id']]['balance']
+                                   + datum['amount'])
+                    note = datum['note']
+                    if note:
+                        try:
+                            date = datetime.datetime.strptime(
+                                note, PARSE_OUTPUT_DATEFORMAT)
+                        except ValueError:
+                            pass
+                        else:
+                            # This is the default case and makes it pretty
+                            note = note_template.format(
+                                amount=money_filter(datum['amount']),
+                                new_balance=money_filter(new_balance),
+                                date=date.strftime(PARSE_OUTPUT_DATEFORMAT))
+                    count += self.core.change_persona_balance(
+                        rs, datum['persona_id'], new_balance,
+                        const.FinanceLogCodes.increase_balance,
+                        change_note=note)
+                    if new_balance >= self.conf["MEMBERSHIP_FEE"]:
+                        code, _, _ = self.change_membership(
+                            rs, datum['persona_id'], is_member=True)
+                        memberships_gained += bool(code)
+                    # Remember the changed balance in case of multiple transfers.
+                    personas[datum['persona_id']]['balance'] = new_balance
+        except psycopg2.extensions.TransactionRollbackError:
+            # We perform a rather big transaction, so serialization errors
+            # could happen.
+            return False, None, None
+        except Exception:
+            # This blanket catching of all exceptions is a last resort. We try
+            # to do enough validation, so that this should never happen, but
+            # an opaque error (as would happen without this) would be rather
+            # frustrating for the users -- hence some extra error handling
+            # here.
+            self.logger.error(glue(
+                ">>>\n>>>\n>>>\n>>> Exception during transfer processing",
+                "<<<\n<<<\n<<<\n<<<"))
+            self.logger.exception("FIRST AS SIMPLE TRACEBACK")
+            self.logger.error("SECOND TRY CGITB")
+            self.cgitb_log()
+            return False, index, None
+        return True, count, memberships_gained
 
     @access("finance_admin")
     def finance_statistics(self, rs: RequestState) -> CdEDBObject:
@@ -897,13 +1044,261 @@ class CdEBackend(AbstractBackend):
                              persona_id=None, change_note=msg)
         return ret
 
+    @access("finance_admin")
+    def process_for_semester_bill(self, rs: RequestState, period_id: int,
+                                  addresscheck: bool, testrun: bool
+                                  ) -> Tuple[bool, Optional[CdEDBObject]]:
+        """Atomized call to bill one persona.
+
+        :returns: A tuple consisting of a boolean signalling whether there
+            is more work to do and an optional persona which is present if
+            work was performed on this invocation.
+        """
+        period_id = affirm(int, period_id)
+        addresscheck = affirm(bool, addresscheck)
+        testrun = affirm(bool, testrun)
+        with Atomizer(rs):
+            period = self.get_period(rs, period_id)
+            persona_id = self.core.next_persona(
+                rs, period['billing_state'], is_member=True, is_archived=False)
+            if testrun:
+                persona_id = rs.user.persona_id
+            # We are finished if we reached the end or if this was previously done.
+            if not persona_id or period['billing_done']:
+                if not period['billing_done']:
+                    self.finish_semester_bill(rs, addresscheck)
+                return False, None
+            period_update = {
+                'id': period_id,
+                'billing_state': persona_id,
+            }
+            persona = self.core.get_cde_user(rs, persona_id)
+            period_update['billing_count'] = period['billing_count'] + 1
+            if not testrun:
+                self.set_period(rs, period_update)
+
+            return True, persona
+
+    @access("finance_admin")
+    def process_for_semester_prearchival(self, rs: RequestState, period_id: int,
+                                         testrun: bool
+                                         ) -> Tuple[bool, Optional[CdEDBObject]]:
+        """Atomized call to warn one persona prior to archival.
+
+        :returns: A tuple consisting of a boolean signalling whether there
+            is more work to do and an optional persona which is present if
+            work was performed on this invocation.
+        """
+        period_id = affirm(int, period_id)
+        testrun = affirm(bool, testrun)
+        with Atomizer(rs):
+            period = self.get_period(rs, period_id)
+            persona_id = self.core.next_persona(
+                rs, period['archival_notification_state'], is_member=None,
+                is_archived=False)
+            if testrun:
+                persona_id = rs.user.persona_id
+            # We are finished if we reached the end or if this was previously done.
+            if not persona_id or period['archival_notification_done']:
+                if not period['archival_notification_done']:
+                    self.finish_archival_notification(rs)
+                return False, None
+            period_update = {
+                'id': period_id,
+                'archival_notification_state': persona_id,
+            }
+            is_archivable = self.core.is_persona_automatically_archivable(
+                rs, persona_id)
+            persona = None
+            if is_archivable or testrun:
+                persona = self.core.get_persona(rs, persona_id)
+                period_update['archival_notification_count'] = \
+                    period['archival_notification_count'] + 1
+            if not testrun:
+                self.set_period(rs, period_update)
+            return True, persona
+
+    @access("finance_admin")
+    def process_for_semester_eject(self, rs: RequestState, period_id: int,
+                                   ) -> Tuple[bool, Optional[CdEDBObject]]:
+        """Atomized call to eject one (soon to be ex-)member.
+
+        :returns: A tuple consisting of a boolean signalling whether there
+            is more work to do and an optional persona which is present if
+            work was performed on this invocation.
+        """
+        period_id = affirm(int, period_id)
+        with Atomizer(rs):
+            period = self.get_period(rs, period_id)
+            persona_id = self.core.next_persona(
+                rs, period['ejection_state'], is_member=True, is_archived=False)
+            # We are finished if we reached the end or if this was previously done.
+            if not persona_id or period['ejection_done']:
+                if not period['ejection_done']:
+                    self.finish_semester_ejection(rs)
+                return False, None
+            period_update = {
+                'id': period_id,
+                'ejection_state': persona_id,
+            }
+            persona = self.core.get_cde_user(rs, persona_id)
+            do_eject = (persona['balance'] < self.conf["MEMBERSHIP_FEE"]
+                        and not persona['trial_member'])
+            if do_eject:
+                self.change_membership(rs, persona_id, is_member=False)
+                period_update['ejection_count'] = \
+                    period['ejection_count'] + 1
+                period_update['ejection_balance'] = \
+                    period['ejection_balance'] + persona['balance']
+            else:
+                persona = None  # type: ignore[assignment]
+            self.set_period(rs, period_update)
+            return True, persona
+
+    @access("finance_admin")
+    def process_for_semester_archival(self, rs: RequestState, period_id: int,
+                                      ) -> Tuple[bool, Optional[CdEDBObject]]:
+        """Atomized call to archive one persona.
+
+        :returns: A tuple consisting of a boolean signalling whether there
+            is more work to do and an optional persona which is present if
+            an error occured during archival (deviating from the common
+            pattern for this functions which return a persona if work was
+            performed on this invocation).
+        """
+        period_id = affirm(int, period_id)
+        with Atomizer(rs):
+            period = self.get_period(rs, period_id)
+            persona_id = self.core.next_persona(
+                rs, period['archival_state'], is_member=False, is_archived=False)
+            # We are finished if we reached the end or if this was previously done.
+            if not persona_id or period['archival_done']:
+                if not period['archival_done']:
+                    self.finish_automated_archival(rs)
+                return False, None
+            period_update = {
+                'id': period_id,
+                'archival_state': persona_id,
+            }
+            persona = None
+            if self.core.is_persona_automatically_archivable(
+                    rs, persona_id, reference_date=period['billing_done']):
+                note = "Autmoatisch archiviert wegen Inaktivität."
+                try:
+                    code = self.core.archive_persona(rs, persona_id, note)
+                except ArchiveError:
+                    self.logger.exception(f"Unexpected error during archival of"
+                                          f" persona {persona_id}.")
+                    persona = {'persona_id': persona_id}
+                else:
+                    if code:
+                        period_update['archival_count'] = \
+                            period['archival_count'] + 1
+                    else:
+                        self.logger.error(
+                            f"Automated archival of persona {persona_id} failed"
+                            f" for unknown reasons.")
+                        persona = {'id': persona_id}
+            self.set_period(rs, period_update)
+            return True, persona
+
+    @access("finance_admin")
+    def process_for_semester_balance(self, rs: RequestState, period_id: int,
+                                     ) -> Tuple[bool, Optional[CdEDBObject]]:
+        """Atomized call to update the balance of one member.
+
+        :returns: A tuple consisting of a boolean signalling whether there
+            is more work to do and an optional persona which is present if
+            work was performed on this invocation.
+        """
+        period_id = affirm(int, period_id)
+        with Atomizer(rs):
+            period = self.get_period(rs, period_id)
+            persona_id = self.core.next_persona(
+                rs, period['balance_state'], is_member=True, is_archived=False)
+            # We are finished if we reached the end or if this was previously done.
+            if not persona_id or period['balance_done']:
+                if not period['balance_done']:
+                    self.finish_semester_balance_update(rs)
+                return False, None
+            persona = self.core.get_cde_user(rs, persona_id)
+            period_update = {
+                'id': period_id,
+                'balance_state': persona_id,
+            }
+            if (persona['balance'] < self.conf["MEMBERSHIP_FEE"]
+                    and not persona['trial_member']):
+                # TODO maybe fail more gracefully here?
+                # Maybe set balance to 0 and send a mail or something.
+                raise ValueError(n_("Balance too low."))
+            else:
+                if persona['trial_member']:
+                    update = {
+                        'id': persona_id,
+                        'trial_member': False,
+                    }
+                    self.core.change_persona(
+                        rs, update,
+                        change_note="Probemitgliedschaft beendet."
+                    )
+                    period_update['balance_trialmembers'] = \
+                        period['balance_trialmembers'] + 1
+                else:
+                    new_b = persona['balance'] - self.conf["MEMBERSHIP_FEE"]
+                    note = "Mitgliedsbeitrag abgebucht ({}).".format(
+                        money_filter(self.conf["MEMBERSHIP_FEE"]))
+                    self.core.change_persona_balance(
+                        rs, persona_id, new_b,
+                        const.FinanceLogCodes.deduct_membership_fee,
+                        change_note=note)
+                    new_total = (period['balance_total']
+                                 + self.conf["MEMBERSHIP_FEE"])
+                    period_update['balance_total'] = new_total
+            self.set_period(rs, period_update)
+            return True, persona
+
+    @access("finance_admin")
+    def process_for_expuls_check(self, rs: RequestState, expuls_id: int,
+                                 testrun: bool) -> Tuple[bool, Optional[CdEDBObject]]:
+        """Atomized call to initiate addres check.
+
+        :returns: A tuple consisting of a boolean signalling whether there
+            is more work to do and an optional persona which is present if
+            work was performed on this invocation.
+        """
+        expuls_id = affirm(int, expuls_id)
+        testrun = affirm(bool, testrun)
+        with Atomizer(rs):
+            expuls = self.get_expuls(rs, expuls_id)
+            persona_id = self.core.next_persona(
+                rs, expuls['addresscheck_state'],
+                is_member=True, is_archived=False)
+            if testrun:
+                persona_id = rs.user.persona_id
+            # We are finished if we reached the end or if this was previously done.
+            if not persona_id or expuls['addresscheck_done']:
+                if not expuls['addresscheck_done']:
+                    self.finish_expuls_addresscheck(
+                        rs, skip=False)
+                return False, None
+            persona = self.core.get_cde_user(rs, persona_id)
+            if not testrun:
+                expuls_update = {
+                    'id': expuls_id,
+                    'addresscheck_state': persona_id,
+                    'addresscheck_count': expuls['addresscheck_count'] + 1,
+                }
+                self.set_expuls(rs, expuls_update)
+            return True, persona
+
     @access("member", "cde_admin")
     def get_member_stats(self, rs: RequestState
                          ) -> Tuple[CdEDBObject, CdEDBObject, CdEDBObject]:
         """Retrieve some generic statistics about members."""
         # Simple stats first.
         query = """SELECT
-            num_members, num_of_searchable, num_of_trial, num_ex_members, num_all
+            num_members, num_of_searchable, num_of_trial, num_of_printed_expuls,
+            num_ex_members, num_all
         FROM
             (
                 SELECT COUNT(*) AS num_members
@@ -921,6 +1316,11 @@ class CdEBackend(AbstractBackend):
                 WHERE is_member = True AND trial_member = True
             ) AS trial_count,
             (
+                SELECT COUNT(*) AS num_of_printed_expuls
+                FROM core.personas
+                WHERE is_member = True and paper_expuls = True
+            ) AS printed_expuls_count,
+            (
                 SELECT COUNT(*) AS num_ex_members
                 FROM core.personas
                 WHERE is_cde_realm = True AND is_member = False
@@ -935,7 +1335,7 @@ class CdEBackend(AbstractBackend):
 
         simple_stats = OrderedDict((k, data[k]) for k in (
             n_("num_members"), n_("num_of_searchable"), n_("num_of_trial"),
-            n_("num_ex_members"), n_("num_all")))
+            n_("num_of_printed_expuls"), n_("num_ex_members"), n_("num_all")))
 
         # TODO: improve this type annotation with a new mypy version.
         def query_stats(select: str, condition: str, order: str, limit: int = 0
@@ -1003,7 +1403,8 @@ class CdEBackend(AbstractBackend):
 
         # Unique event attendees per year:
         query = """SELECT
-            COUNT(DISTINCT persona_id) AS num, EXTRACT(year FROM events.tempus)::integer AS datum
+            COUNT(DISTINCT persona_id) AS num,
+            EXTRACT(year FROM events.tempus)::integer AS datum
         FROM
             (
                 past_event.institutions
@@ -1025,6 +1426,174 @@ class CdEBackend(AbstractBackend):
             (e['datum'], e['num']) for e in self.query_all(rs, query, ()))
 
         return simple_stats, other_stats, year_stats
+
+    def _perform_one_batch_admission(self, rs: RequestState, datum: CdEDBObject,
+                                     trial_membership: bool, consent: bool
+                                     ) -> Optional[bool]:
+        """Uninlined code from perform_batch_admission().
+
+        :returns: None if nothing happened. True if a new account was created. False if
+            an existing account was modified.
+        """
+        ret = False
+        batch_fields = (
+            'family_name', 'given_names', 'display_name', 'title', 'name_supplement',
+            'birth_name', 'gender', 'address_supplement', 'address',
+            'postal_code', 'location', 'country', 'telephone',
+            'mobile', 'birthday')  # email omitted as it is handled separately
+        if datum['resolution'] == LineResolutions.skip:
+            return None
+        elif datum['resolution'] == LineResolutions.create:
+            new_persona = copy.deepcopy(datum['persona'])
+            new_persona.update({
+                'is_member': True,
+                'trial_member': trial_membership,
+                'paper_expuls': True,
+                'is_searchable': consent,
+            })
+            persona_id = self.core.create_persona(rs, new_persona)
+            ret = True
+        elif datum['resolution'].is_modification():
+            persona_id = datum['doppelganger_id']
+            current = self.core.get_persona(rs, persona_id)
+            if current['is_archived']:
+                if current['is_purged']:
+                    raise RuntimeError(n_("Cannot restore purged account."))
+                self.core.dearchive_persona(
+                    rs, persona_id, datum['persona']['username'])
+                current['username'] = datum['persona']['username']
+            if datum['update_username']:
+                if current['username'] != datum['persona']['username']:
+                    self.core.change_username(
+                        rs, persona_id, datum['persona']['username'], password=None)
+            if not current['is_cde_realm']:
+                # Promote to cde realm dependent on current realm
+                promotion: CdEDBObject = {
+                    field: None for field in CDE_TRANSITION_FIELDS}
+                # The ream independent upgrades of the persona. They are applied at last
+                # to prevent unintentional overrides
+                upgrades = {
+                    'is_cde_realm': True,
+                    'is_event_realm': True,
+                    'is_assembly_realm': True,
+                    'is_ml_realm': True,
+                    'decided_search': False,
+                    'trial_member': False,
+                    'paper_expuls': True,
+                    'bub_search': False,
+                    'id': persona_id,
+                }
+                # This applies a part of the newly imported data necessary for realm
+                # transition. The remaining data will be updated later.
+                mandatory_fields = {
+                    field for field, validator in CDE_TRANSITION_FIELDS.items()
+                    if field not in upgrades and not is_optional(validator)
+                }
+                assert mandatory_fields <= set(batch_fields)
+                # It is pure incident that only event users have additional (optional)
+                # data they share with cde users and which must be honoured during realm
+                # transition. This may be changed if a new user tier is introduced.
+                if not current['is_event_realm']:
+                    if not datum['resolution'].do_update():
+                        raise RuntimeError(n_("Need extra data."))
+                    for field in mandatory_fields:
+                        promotion[field] = datum['persona'][field]
+                else:
+                    current = self.core.get_event_user(rs, persona_id)
+                    # take care that we do not override existent data
+                    current_fields = {
+                        field for field in CDE_TRANSITION_FIELDS
+                        if current.get(field) is not None
+                    }
+                    for field in current_fields:
+                        promotion[field] = current[field]
+                    for field in mandatory_fields:
+                        if promotion[field] is None:
+                            promotion[field] = datum['persona'][field]
+                # apply the actual changes
+                promotion.update(upgrades)
+                self.core.change_persona_realms(
+                    rs, promotion, change_note="Datenübernahme nach Massenaufnahme")
+            if datum['resolution'].do_trial():
+                self.change_membership(
+                    rs, datum['doppelganger_id'], is_member=True)
+                update = {
+                    'id': datum['doppelganger_id'],
+                    'trial_member': True,
+                }
+                self.core.change_persona(
+                    rs, update, may_wait=False,
+                    change_note="Probemitgliedschaft erneuert.")
+            if datum['resolution'].do_update():
+                update = {'id': datum['doppelganger_id']}
+                for field in batch_fields:
+                    update[field] = datum['persona'][field]
+                self.core.change_persona(
+                    rs, update, may_wait=True, force_review=True,
+                    change_note="Import aktualisierter Daten.")
+        else:
+            raise RuntimeError(n_("Impossible."))
+        if datum['pevent_id'] and persona_id:
+            self.pastevent.add_participant(
+                rs, datum['pevent_id'], datum['pcourse_id'], persona_id,
+                is_instructor=datum['is_instructor'], is_orga=datum['is_orga'])
+        return ret
+
+    @access("cde_admin")
+    def perform_batch_admission(self, rs: RequestState, data: List[CdEDBObject],
+                                trial_membership: bool, consent: bool
+                                ) -> Tuple[bool, Optional[int], Optional[int]]:
+        """Atomized call to recruit new members.
+
+        The frontend wants to do this in its entirety or not at all, so this
+        needs to be in atomized context.
+
+        :returns: A tuple consisting of a bool and two optional integers.:
+            A boolean signalling success.
+            If the operation was successful:
+                An integer signalling the number of newly created accounts.
+                An integer signalling the number of modified/renewed accounts.
+            If the operation was not successful:
+                If a TransactionRollbackError occrured:
+                    Both integer parameters are None.
+                Otherwise:
+                    The index where the error occured.
+                    The second parameter is None.
+        """
+        data = affirm_array(vtypes.BatchAdmissionEntry, data)
+        trial_membership = affirm(bool, trial_membership)
+        consent = affirm(bool, consent)
+        # noinspection PyBroadException
+        try:
+            with Atomizer(rs):
+                count_new = count_renewed = 0
+                for index, datum in enumerate(data, start=1):
+                    account_created = self._perform_one_batch_admission(
+                        rs, datum, trial_membership, consent)
+                    if account_created is None:
+                        pass
+                    elif account_created:
+                        count_new += 1
+                    else:
+                        count_renewed += 1
+        except psycopg2.extensions.TransactionRollbackError:
+            # We perform a rather big transaction, so serialization errors
+            # could happen.
+            return False, None, None
+        except Exception:
+            # This blanket catching of all exceptions is a last resort. We try
+            # to do enough validation, so that this should never happen, but
+            # an opaque error (as would happen without this) would be rather
+            # frustrating for the users -- hence some extra error handling
+            # here.
+            self.logger.error(glue(
+                ">>>\n>>>\n>>>\n>>> Exception during batch creation",
+                "<<<\n<<<\n<<<\n<<<"))
+            self.logger.exception("FIRST AS SIMPLE TRACEBACK")
+            self.logger.error("SECOND TRY CGITB")
+            self.cgitb_log()
+            return False, index, None
+        return True, count_new, count_renewed
 
     @access("searchable", "core_admin", "cde_admin")
     def submit_general_query(self, rs: RequestState,
