@@ -17,6 +17,8 @@ functions:
   a cleanup_protected state. This function may be temporarily turned off for a given
   object without issue.
 """
+# We allow 120 line length here.
+# pylint: disable=line-too-long
 
 from typing import AbstractSet, Collection, Optional
 
@@ -42,18 +44,19 @@ class SubscriptionManager:
 
     def __init__(
         self, *,
-        error_matrix: ActionStateErrorMatrix = SUBSCRIPTION_ERROR_MATRIX,
+        error_matrix: ActionStateErrorMatrix = None,
         unwritten_states: Optional[StateColl] = None,
     ) -> None:
         """
         Create an instance of a `SubscriptionManager`.
 
-        :param error_matrix: You may provide an alternative matrix defining the
+        :param error_matrix: Defaults to `subman.machine.SUBSCRIPTION_ERROR_MATRIX`.
+            You may provide an alternative matrix defining the
             relations between actions and states.
         :param unwritten_states: Provide this if you want to keep track of a subset of
             all `SubscriptionState`s, that should not be written to your database.
         """
-        self.error_matrix = error_matrix
+        self.error_matrix = error_matrix or SUBSCRIPTION_ERROR_MATRIX
         self.unwritten_states: StateSet = set(unwritten_states or ())
         if (self.unwritten_states &
                 {SubscriptionState.subscribed, SubscriptionState.unsubscribed}):
@@ -91,9 +94,7 @@ class SubscriptionManager:
         of them may be appropriate to use from the current state.
         """
 
-        # TODO: `if exception := self.get_error(action, old_state)`.
-        exception = self._get_error(action, old_state)
-        if exception:
+        if exception := self._get_error(action, old_state):
             raise exception
 
     @staticmethod
