@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+# pylint: disable=missing-module-docstring
 
 import copy
 import datetime
@@ -509,14 +510,15 @@ class TestCoreBackend(BackendTest):
                 if persona[key] is None:
                     persona[key] = True
         merge_dicts(data, persona)
-        self.assertLess(0, self.core.change_persona_realms(self.key, data))
+        change_note = "Bereichsänderung"
+        self.assertLess(0, self.core.change_persona_realms(self.key, data, change_note))
         log_entry = {
             'id': 1001,
             'ctime': nearly_now(),
             'code': const.CoreLogCodes.realm_change,
             'persona_id': persona_id,
             'submitted_by': self.user['id'],
-            'change_note': 'Bereiche geändert.'
+            'change_note': change_note,
         }
         _, expected_log = self.core.retrieve_log(self.key)
         self.assertIn(log_entry, expected_log)
