@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+# pylint: disable=protected-access,missing-module-docstring
 
 from typing import Collection, Set, cast
 
@@ -43,7 +44,8 @@ class TestMlBackend(BackendTest):
             self.ml.merge_accounts(self.key,
                                    source_persona_id=USER_DICT['nina']['id'],
                                    target_persona_id=berta_id)
-        self.assertEqual("Source User is admin and can not be merged.", str(e.exception))
+        self.assertEqual("Source User is admin and can not be merged.",
+                         str(e.exception))
 
         with self.assertRaises(ValueError) as e:
             self.ml.merge_accounts(self.key,
@@ -340,7 +342,7 @@ class TestMlBackend(BackendTest):
             'title': 'Proletarier aller Länder',
         }
         self.assertLess(0, self.ml.create_mailinglist(self.key, new_data))
-        assert isinstance(new_data['moderators'], Set)
+        assert isinstance(new_data['moderators'], Set)  # pylint: disable=isinstance-second-argument-not-valid-type
         new_data['moderators'] |= {100000}
         with self.assertRaises(ValueError):
             self.ml.create_mailinglist(self.key, new_data)
@@ -626,7 +628,8 @@ class TestMlBackend(BackendTest):
                          state=SS.unsubscribed, kind="info")
 
         # This does some basic override testing.
-        self._change_sub(self.user['id'], mailinglist_id, SA.add_unsubscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.add_unsubscription_override,
                          state=SS.unsubscription_override)
         self._change_sub(self.user['id'], mailinglist_id, SA.subscribe,
                          state=SS.unsubscription_override, kind="error")
@@ -639,8 +642,8 @@ class TestMlBackend(BackendTest):
         # transition between non-core states are forbidden
         self._change_sub(self.user['id'], mailinglist_id, SA.add_subscription_override,
                          state=SS.unsubscription_override, kind="error")
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_unsubscription_override,
-                         state=SS.unsubscribed)
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_unsubscription_override, state=SS.unsubscribed)
         self._change_sub(self.user['id'], mailinglist_id, SA.add_subscription_override,
                          state=SS.subscription_override)
         self._change_sub(self.user['id'], mailinglist_id, SA.remove_subscriber,
@@ -651,7 +654,8 @@ class TestMlBackend(BackendTest):
                          state=SS.subscription_override, kind="info")
         self._change_sub(self.user['id'], mailinglist_id, SA.unsubscribe,
                          state=SS.unsubscribed)
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_subscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_subscription_override,
                          state=SS.unsubscribed, kind="error")
 
         # You cannot request subscriptions to such lists
@@ -670,21 +674,25 @@ class TestMlBackend(BackendTest):
                          state=SS.unsubscribed, kind="info")
 
         # This does more override management testing
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_unsubscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_unsubscription_override,
                          state=SS.unsubscribed, kind="error")
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_subscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_subscription_override,
                          state=SS.unsubscribed, kind="error")
         self._change_sub(self.user['id'], mailinglist_id, SA.subscribe,
                          state=SS.subscribed)
-        self._change_sub(self.user['id'], mailinglist_id, SA.add_unsubscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.add_unsubscription_override,
                          state=SS.unsubscription_override)
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_unsubscription_override,
-                         state=SS.unsubscribed)
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_unsubscription_override, state=SS.unsubscribed)
         self._change_sub(self.user['id'], mailinglist_id, SA.add_subscription_override,
                          state=SS.subscription_override)
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_subscription_override,
-                         state=SS.subscribed)
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_subscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_subscription_override, state=SS.subscribed)
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_subscription_override,
                          state=SS.subscribed, kind="error")
 
         # This tests the unsubscription reset.
@@ -814,14 +822,15 @@ class TestMlBackend(BackendTest):
                          state=SS.unsubscribed)
 
         # Test blocks
-        self._change_sub(self.user['id'], mailinglist_id, SA.add_unsubscription_override,
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.add_unsubscription_override,
                          state=SS.unsubscription_override)
         self._change_sub(self.user['id'], mailinglist_id, SA.subscribe,
                          state=SS.unsubscription_override, kind="error")
         self._change_sub(self.user['id'], mailinglist_id, SA.add_subscriber,
                          state=SS.unsubscription_override, kind="error")
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_unsubscription_override,
-                         state=SS.unsubscribed)
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_unsubscription_override, state=SS.unsubscribed)
 
         # Test forced subscriptions
         self._change_sub(self.user['id'], mailinglist_id, SA.subscribe,
@@ -925,8 +934,8 @@ class TestMlBackend(BackendTest):
                          state=SS.subscription_override, kind="error")
 
         # Remove subscription
-        self._change_sub(self.user['id'], mailinglist_id, SA.remove_subscription_override,
-                         state=SS.subscribed)
+        self._change_sub(self.user['id'], mailinglist_id,
+                         SA.remove_subscription_override, state=SS.subscribed)
 
         # Cron testing
         self.ml.write_subscription_states(self.key, (mailinglist_id,))
