@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+# pylint: disable=missing-module-docstring
 
 import io
 import os
@@ -51,7 +52,7 @@ class TestScript(unittest.TestCase):
         core = make_backend("core", proxy=False)
         self.assertTrue(isinstance(core, CoreBackend))
         coreproxy = make_backend("core", proxy=True)
-        self.assertEqual(coreproxy._get_backend_class(), CoreBackend)
+        self.assertEqual(coreproxy.get_backend_class(), CoreBackend)
         self.assertEqual(
             make_backend("core", proxy=False, LOCKDOWN=42).conf["LOCKDOWN"], 42)
         # This way of writing to a tempporary file mirrors exactly what happens
